+++ gplugins/sax/parameter.py
@@ -140,8 +140,8 @@ class LithoParameter(Parameter):
 
     def layer_dilation_erosion(self, component, dilation_value):
         temp_component = gf.Component()
-        for layer, layer_polygons in component.get_polygons(by_spec=True).items():
-            if layer == self.layer:
+        for layer, layer_polygons in component.get_polygons_points().items():
+            if layer == self.layer.layer1.layer1.layer:
                 # Make sure all layer polygons are fused properly
                 shapely_polygons = [
                     shapely.geometry.Polygon(polygon) for polygon in layer_polygons
@@ -161,15 +161,16 @@ class LithoParameter(Parameter):
                         if hasattr(buffered_polygons, "geoms")
                         else [buffered_polygons]
                     ):
-                        temp_component.add_polygon(
-                            buffered_polygon.exterior.coords, layer=layer
+                        coords_tuple = tuple(
+                            map(tuple, buffered_polygon.exterior.coords)
                         )
+                        temp_component.add_polygon(coords_tuple, layer=layer)
             else:
-                for layer_polygon in layer_polygons:
-                    temp_component.add_polygon(layer_polygon, layer=layer)
+                for poly in layer_polygons:
+                    temp_component.add_polygon(poly, layer=layer)
         # Transform ports
         ports = []
-        for port in component.get_ports():
+        for port in component.ports:
             if port.layer == self.layer:
                 port.width += 2 * dilation_value
                 old_center_x, old_center_y = port.center
@@ -186,21 +187,21 @@ class LithoParameter(Parameter):
 
     def layer_x_offset(self, component, offset_value):
         temp_component = gf.Component()
-        for layer, layer_polygons in component.get_polygons(by_spec=True).items():
+        for layer, layer_polygons in component.get_polygons_points().items():
             for layer_polygon in layer_polygons:
-                if layer == self.layer:
+                if layer == self.layer.layer1.layer1.layer:
                     shapely_polygon = shapely.geometry.Polygon(layer_polygon)
                     translated_polygon = translate(
                         shapely_polygon, xoff=offset_value, yoff=0.0
                     )
-                    temp_component.add_polygon(
-                        translated_polygon.exterior.coords, layer=layer
-                    )
+
+                    coords_tuple = tuple(map(tuple, translated_polygon.exterior.coords))
+                    temp_component.add_polygon(coords_tuple, layer=layer)
                 else:
                     temp_component.add_polygon(layer_polygon, layer=layer)
         # Transform ports
         ports = []
-        for port in component.get_ports():
+        for port in component.ports:
             if port.layer == self.layer:
                 old_center_x, old_center_y = port.center
                 new_center_x = old_center_x + offset_value
@@ -211,21 +212,20 @@ class LithoParameter(Parameter):
 
     def layer_y_offset(self, component, offset_value):
         temp_component = gf.Component()
-        for layer, layer_polygons in component.get_polygons(by_spec=True).items():
+        for layer, layer_polygons in component.get_polygons_points().items():
             for layer_polygon in layer_polygons:
-                if layer == self.layer:
+                if layer == self.layer.layer1.layer1.layer:
                     shapely_polygon = shapely.geometry.Polygon(layer_polygon)
                     translated_polygon = translate(
                         shapely_polygon, xoff=0.0, yoff=offset_value
                     )
-                    temp_component.add_polygon(
-                        translated_polygon.exterior.coords, layer=layer
-                    )
+                    coords = tuple(map(tuple, translated_polygon.exterior.coords))
+                    temp_component.add_polygon(coords, layer=layer)
                 else:
                     temp_component.add_polygon(layer_polygon, layer=layer)
         # Transform ports
         ports = []
-        for port in component.get_ports():
+        for port in component.ports:
             if port.layer == self.layer:
                 old_center_x, old_center_y = port.center
                 new_center_y = old_center_y + offset_value
@@ -236,8 +236,8 @@ class LithoParameter(Parameter):
 
     def layer_round_corners(self, component, round_value):
         temp_component = gf.Component()
-        for layer, layer_polygons in component.get_polygons(by_spec=True).items():
-            if layer == self.layer:
+        for layer, layer_polygons in component.get_polygons_points().items():
+            if layer == self.layer.layer1.layer1.layer:
                 # Make sure all layer polygons are fused properly
                 shapely_polygons = [
                     shapely.geometry.Polygon(polygon) for polygon in layer_polygons
@@ -254,15 +254,14 @@ class LithoParameter(Parameter):
                         .buffer(-2 * round_value, join_style=1)
                         .buffer(round_value, join_style=1)
                     )
-                    temp_component.add_polygon(
-                        buffered_polygon.exterior.coords, layer=layer
-                    )
+                    coords = tuple(map(tuple, buffered_polygon.exterior.coords))
+                    temp_component.add_polygon(coords, layer=layer)
             else:
                 for layer_polygon in layer_polygons:
                     temp_component.add_polygon(layer_polygon, layer=layer)
         # Transform ports
         ports = []
-        for port in component.get_ports():
+        for port in component.ports:
             if port.layer == self.layer:
                 # Patch port
                 patch_polygon_x1, patch_polygon_y1 = port.center
@@ -330,24 +329,24 @@ if __name__ == "__main__":
     )
     c.add_port(name="o1", center=(0, 1), width=1, orientation=0, layer=1)
     c.add_port(name="o2", center=(3, -2), width=1, orientation=90, layer=1)
-    c.show(show_ports=True)
+    c.show()
 
-    param = LithoParameter(layername="core")
-    eroded_c = param.layer_dilation_erosion(c, 0.2)
-    eroded_c.show(show_ports=True)
+    # param = LithoParameter(layername="core")
+    # eroded_c = param.layer_dilation_erosion(c, 0.2)
+    # eroded_c.show()
 
-    param = LithoParameter(layername="core")
-    eroded_c = param.layer_dilation_erosion(c, -0.2)
-    eroded_c.show(show_ports=True)
+    # param = LithoParameter(layername="core")  # FIXME
+    # eroded_c = param.layer_dilation_erosion(c, -0.2)
+    # eroded_c.show()
 
-    param = LithoParameter(layername="core")
-    eroded_c = param.layer_x_offset(c, 0.5)
-    eroded_c.show(show_ports=True)
+    # param = LithoParameter(layername="core")
+    # eroded_c = param.layer_x_offset(c, 0.5)
+    # eroded_c.show()
 
-    param = LithoParameter(layername="core")
-    eroded_c = param.layer_y_offset(c, 0.5)
-    eroded_c.show(show_ports=True)
+    # param = LithoParameter(layername="core")
+    # eroded_c = param.layer_y_offset(c, 0.5)
+    # eroded_c.show()
 
     param = LithoParameter(layername="core")
     eroded_c = param.layer_round_corners(c, 0.2)
-    eroded_c.show(show_ports=True)
+    eroded_c.show()